requirements.txt
aiogram==2.14.3
aiohttp==3.7.4.post0
appdirs==1.4.4
astroid==2.7.2
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
black==21.7b0
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
emoji==1.4.2
idna==3.2
isort==5.9.3
lazy-object-proxy==1.6.0
mccabe==0.6.1
multidict==5.1.0
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.2.0
pylint==2.10.2
python-decouple==3.4
pytz==2021.1
regex==2021.8.21
toml==0.10.2
tomli==1.2.1
typing-extensions==3.10.0.0
wrapt==1.12.1
yarl==1.6.3
